refactor(embed): move embed files to internal package
