+++ .gitignore
@@ -102,3 +102,4 @@ tools-stamp
 sonr.log
 
 interchaintest-downloader
+.haptic
