+++ internal/embed/main.js
