+++ internal/embed/sw.js
