+++ internal/embed/utils.go
