+++ internal/embed/webworker.go
