+++ pkg/gateway/handlers/register_handler.go
@@ -3,21 +3,20 @@ package handlers
 import (
 	"github.com/labstack/echo/v4"
 	"github.com/onsonr/sonr/pkg/gateway/middleware"
-	"github.com/onsonr/sonr/pkg/gateway/types"
 	"github.com/onsonr/sonr/pkg/gateway/views"
 )
 
 func RenderProfileCreate(c echo.Context) error {
 	// numF, numL := middleware.GetHumanVerificationNumbers(c)
-	params := types.CreateProfileParams{
+	params := middleware.CreateProfileParams{
 		FirstNumber: int(middleware.CurrentBlock(c)),
 		LastNumber:  int(middleware.CurrentBlock(c)),
 	}
-	return middleware.Render(c, views.RegisterProfileView(params))
+	return middleware.Render(c, views.RegisterProfileView(params.FirstNumber, params.LastNumber))
 }
 
 func RenderPasskeyCreate(c echo.Context) error {
-	return middleware.Render(c, views.RegisterPasskeyView(types.CreatePasskeyParams{}))
+	return middleware.Render(c, views.RegisterPasskeyView("", "", "", "", ""))
 }
 
 func RenderVaultLoading(c echo.Context) error {
